+++ src/routes/tasks/components/TaskDrawer.tsx
@@ -207,7 +207,6 @@ const TaskDrawer = ({
                                                 isTruncated
                                                 maxWidth="370px"
                                                 textAlign="right"
-                                                marginLeft="auto"
                                             >
                                                 <Link
                                                     href={compilePath(