+++ CHANGELOG.md
@@ -64,6 +64,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   General improvements of the perf browser component (#279)
 -   Added ellipsis to too long links (#274)
 -   New logo (#281)
+-   Align TaskDrawer children to the left (#305)
 
 ## [0.6.0] - 2021-12-01